+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [Refactor] Migrate file manager to decompose
 - [Refactor] Migrate options screen to decompose
+- [Refactor] Migrate device screen to decompose
 - [Feature] New report bug
 
 # 1.6.7